(pd) substep cleanup
* factor out formProcessing.js
* consolidate basic substeps
