+++ protocol-designer/src/steplist/generateSubsteps.js
@@ -1,293 +1,94 @@
 // @flow
-import {humanize} from '../utils'
+import mapValues from 'lodash/mapValues'
+import range from 'lodash/range'
+
+import {
+  formHasErrors,
+  type ValidFormAndErrors
+} from './formProcessing'
 
 import type {
-  StepType,
   StepIdType,
-  FormData,
-  ProcessedFormData,
-  TransferForm,
-  ConsolidateForm,
-  TransferFormData
+  SubSteps,
+  PauseFormData
 } from './types'
 
-import type {ConsolidateFormData} from '../step-generation'
+function _transferSubsteps (form: *, stepId: StepIdType) {
+  const {
+    sourceWells,
+    destWells
+    // sourceLabware, // TODO: show labware & volume, see new designs
+    // destLabware,
+    // volume
+  } = form
 
-// TODO LATER Ian 2018-03-01 remove or consolidate these 2 similar types?
-export type ValidFormAndErrors = {
-  errors: {[string]: string},
-  validatedForm: ProcessedFormData | null // TODO: incompleteData field when this is null?
+  return {
+    stepType: 'transfer',
+    parentStepId: stepId,
+    // TODO Ian 2018-03-02 break up steps when pipette too small
+    rows: range(sourceWells.length).map(i => ({
+      substepId: i,
+      sourceIngredientName: 'ING1', // TODO get ingredients for source/dest wells
+      destIngredientName: 'ING2',
+      sourceWell: sourceWells[i],
+      destWell: destWells[i]
+    }))
+  }
 }
 
-type ValidationAndErrors<F> = {
-  errors: {[string]: string},
-  validatedForm: F | null
-}
+function _consolidateSubsteps (form: *, stepId: StepIdType) {
+  const {
+    sourceWells,
+    destWell
+  } = form
 
-export const generateNewForm = (stepId: StepIdType, stepType: StepType) => {
-  // Add default values to a new step form
-  const baseForm = {
-    id: stepId,
-    stepType: stepType,
-    'step-name': humanize(stepType) + ' ' + (stepId + 1),
-    'step-details': ''
-  }
-
-  if (stepType === 'transfer' || stepType === 'consolidate') {
-    return {
-      ...baseForm,
-      'aspirate--change-tip': 'once'
-    }
-  }
-  if (stepType !== 'pause') {
-    console.warn('generateNewForm: Only transfer, consolidate, & pause forms are supported now. TODO. Got ' + stepType)
-  }
-  return baseForm
-}
-
-export function formHasErrors (form: {errors: {[string]: string}}): boolean {
-  return Object.values(form.errors).length > 0
-}
-
-function _vapTransfer (formData: TransferForm): ValidationAndErrors<TransferFormData> {
-  // TODO when transfer is supported in step-generation,
-  // combine this with consolidate since args are similar
-  // and clean up the parsing errors
-  const pipette = formData['aspirate--pipette']
-  const sourceWells = formData['aspirate--wells'] ? formData['aspirate--wells'].split(',') : []
-  const destWells = formData['dispense--wells'] ? formData['dispense--wells'].split(',') : []
-  const sourceLabware = formData['aspirate--labware']
-  const destLabware = formData['dispense--labware']
-
-  const volume = parseFloat(formData['dispense--volume'])
-
-  const requiredFieldErrors = [
-    'aspirate--pipette',
-    'aspirate--labware',
-    'dispense--labware'
-  ].reduce((acc: {}, fieldName: string) => (!formData[fieldName])
-    ? {...acc, [fieldName]: 'This field is required'}
-    : acc,
-  {})
-
-  // Conditionally add error fields
-  let errors = {...requiredFieldErrors}
-
-  if (!(volume > 0)) {
-    const field = formData.stepType === 'transfer' ? 'dispense--volume' : 'aspirate--volume'
-    errors[field] = 'Volume must be a positive number'
-  }
-
-  if (formData.stepType === 'transfer') {
-    if (sourceWells.length !== destWells.length || sourceWells.length === 0) {
-      errors._mismatchedWells = 'Numbers of wells must match'
-    }
+  const destWellSubstep = {
+    destWell,
+    sourceIngredientName: 'ING1',
+    destIngredientName: 'ING2'
   }
 
   return {
-    errors,
-    validatedForm: (
-      !formHasErrors({errors}) &&
-      // extra explicit for flow
-      pipette &&
-      sourceLabware &&
-      destLabware
-    )
-      ? {
-        stepType: formData.stepType,
-        pipette,
-        sourceWells,
-        destWells,
-        sourceLabware,
-        destLabware,
-        volume
-      }
-      : null
+    stepType: 'consolidate',
+    parentStepId: stepId,
+    rows: [
+      ...sourceWells.map((sourceWell, i) => ({
+        substepId: i,
+        sourceWell: sourceWell,
+        sourceIngredientName: 'ING1'
+      })),
+      destWellSubstep
+    ]
   }
 }
 
-function _vapConsolidate (formData: ConsolidateForm): ValidationAndErrors<ConsolidateFormData> {
-  const pipette = formData['aspirate--pipette']
-  const sourceWells = formData['aspirate--wells'] ? formData['aspirate--wells'].split(',') : []
-  const destWells = formData['dispense--wells'] ? formData['dispense--wells'].split(',') : []
-  const sourceLabware = formData['aspirate--labware']
-  const destLabware = formData['dispense--labware']
-
-  const volume = parseFloat(formData['aspirate--volume'])
-
-  const requiredFieldErrors = [
-    'aspirate--pipette',
-    'aspirate--labware',
-    'dispense--labware'
-  ].reduce((acc, fieldName) => (!formData[fieldName])
-    ? {...acc, [fieldName]: 'This field is required'}
-    : acc,
-  {})
-
-  // Conditionally add error fields
-  let errors = {...requiredFieldErrors}
-
-  if (!(volume > 0)) {
-    errors['aspirate--volume'] = 'Volume must be a positive number'
-  }
-
-  if (sourceWells.length <= 1 || destWells.length !== 1) {
-    errors._mismatchedWells = 'Multiple sources well and exactly one destination well is required.'
-  }
-
-  const blowout = formData['dispense--blowout--labware']
-
-  const delayAfterDispense = formData['dispense--delay--checkbox']
-    ? ((parseFloat(formData['dispense--delay-minutes']) || 0) * 60) +
-      (parseFloat(formData['dispense--delay-seconds'] || 0))
-    : null
-
-  // const disposalVolume =
-
-  const mixFirstAspirate = formData['aspirate--mix--checkbox']
-    ? {
-      volume: parseFloat(formData['aspirate--mix--volume']),
-      times: parseInt(formData['aspirate--mix--time']) // TODO handle unparseable
+export function generateSubsteps (validatedForms: {[StepIdType]: ValidFormAndErrors}): SubSteps {
+  return mapValues(validatedForms, (valForm: ValidFormAndErrors, stepId: StepIdType) => {
+    // Don't try to render with errors. TODO LATER: presentational error state of substeps?
+    if (valForm.validatedForm === null || formHasErrors(valForm)) {
+      return null
     }
-    : null
 
-  // TODO general mix args creator
-  const mixInDestination = formData['dispense--mix--checkbox']
-    ? {
-      volume: parseFloat(formData['dispense--mix--volume']),
-      times: parseInt(formData['dispense--mix--times'])
+    if (valForm.validatedForm.stepType === 'deck-setup') {
+      // No substeps for Deck Setup
+      return null
     }
-    : null
 
-  const disposalVolume = formData['aspirate--disposal-vol--checkbox']
-    ? parseFloat('aspirate--disposal-vol--volume') // TODO handle unparseable
-    : null
+    if (valForm.validatedForm.stepType === 'transfer') {
+      return _transferSubsteps(valForm.validatedForm, stepId)
+    }
 
-  const changeTip = formData['aspirate--change-tip'] || 'always'
-  // It's radiobutton, so one should always be selected.
-  // TODO use default from importable const DEFAULT_CHANGE_TIP_OPTION
+    if (valForm.validatedForm.stepType === 'pause') {
+      // just returns formData
+      const formData: PauseFormData = valForm.validatedForm
+      return formData
+    }
 
-  return {
-    errors,
-    validatedForm: (
-      !formHasErrors({errors}) &&
-      // extra explicit for flow
-      pipette &&
-      sourceLabware &&
-      destLabware
-    )
-      ? {
-        stepType: formData.stepType,
-        pipette,
-        sourceWells,
-        destWell: destWells[0],
-        sourceLabware,
-        destLabware,
-        volume,
-        blowout, // TODO allow user to blowout
-        changeTip,
-        delayAfterDispense,
-        description: 'description would be here 2018-03-01', // TODO get from form
-        mixFirstAspirate,
-        disposalVolume,
-        mixInDestination,
-        preWetTip: formData['aspirate--pre-wet-tip'] || false,
-        touchTipAfterAspirate: formData['aspirate--touch-tip'] || false,
-        touchTipAfterDispense: false, // TODO Ian 2018-03-01 Not in form
-        name: `Consolidate ${formData.id}` // TODO real name for steps
-      }
-      : null
-  }
+    if (valForm.validatedForm.stepType === 'consolidate') {
+      return _consolidateSubsteps(valForm.validatedForm, stepId)
+    }
+
+    console.warn('allSubsteps doesnt support step type: ', valForm.validatedForm.stepType, stepId)
+    return null
+  })
 }
-
-export function validateAndProcessForm (formData: FormData): * { // ValidFormAndErrors
-  if (formData.stepType === 'transfer') {
-    return _vapTransfer(formData)
-  }
-
-  if (formData.stepType === 'consolidate') {
-    return _vapConsolidate(formData)
-  }
-
-  if (formData.stepType === 'pause') {
-    const hours = parseFloat(formData['pause-hour']) || 0
-    const minutes = parseFloat(formData['pause-minute']) || 0
-    const seconds = parseFloat(formData['pause-second']) || 0
-    const totalSeconds = hours * 360 + minutes * 60 + seconds
-
-    const message = formData['pause-message'] || ''
-
-    const errors = {
-      ...(!formData['pause-for-amount-of-time']
-        ? {'pause-for-amount-of-time': 'Pause for amount of time vs pause until user input is required'}
-        : {}
-      ),
-      ...(formData['pause-for-amount-of-time'] === 'true' && (totalSeconds <= 0)
-        ? {'_pause-times': 'Must include hours, minutes, or seconds'}
-        : {}
-      )
-    }
-
-    return {
-      errors,
-      validatedForm: formHasErrors({errors})
-        ? null
-        : {
-          stepType: formData.stepType,
-          waitForUserInput: formData['pause-for-amount-of-time'] === 'false',
-          totalSeconds,
-          hours,
-          minutes,
-          seconds,
-          message
-        }
-    }
-  }
-
-  // Fallback for unsupported step type. Should be unreachable (...right?)
-  return {
-    errors: {
-      '_form': 'Unsupported step type: ' + formData.stepType
-    },
-    validatedForm: null
-  }
-}
-
-// TODO Ian 2018-03-01 -- this is the old transfer commands / subcommands - DELETE!
-// export function generateCommands (data: ProcessedFormData): Array<Command> {
-//   if (data.stepType === 'transfer') {
-//     // TODO: this should be done in validation/preprocessing step
-//     const {sourceWells, destWells, volume, pipette, sourceLabware, destLabware} = data
-//
-//     if (sourceWells.length !== destWells.length) {
-//       throw new Error('generateSubsteps expected matching N:N source:dest wells for transfer')
-//     }
-//
-//     // TODO strings should already be numbers, validation should already happen
-//     // TODO handle touch tip
-//     return flatMap(zip(sourceWells, destWells), ([sourceWell, destWell]): Array<Command> => [
-//       {
-//         commandType: 'aspirate',
-//         volume, // TODO add dispense volume etc
-//         pipette,
-//         labware: sourceLabware,
-//         well: sourceWell
-//       },
-//       {
-//         commandType: 'dispense',
-//         volume,
-//         pipette,
-//         labware: destLabware,
-//         well: destWell
-//       }
-//     ])
-//   }
-//   // TODO IMMEDIATELY Ian 2018-02-14 why does this keep getting undefined stepType?
-//   console.warn('generateCommands only supports transfer, got: ' + data.stepType, data)
-//   return [] // TODO
-// }
-
-// export function generateSubsteps (): Array<StepSubItemData> {
-//   // TODO: create substeps from formData. It doesn't show all the aspirate/dispenses because it ignores mix, etc.
-//   return {}
-// }
